+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Next Release
+## v7.9.0 (2023-01-18)
 
 - Adds `all` function to `Pickup` to retrieve all pickups
 - Adds `retrieve_payload` and `retrieve_all_payloads` functions to `Event`